Cast to prevent overflow

+++ src/bef.c
@@ -257,7 +257,7 @@ static void bef_openfec_destroy(of_session_t *session)
 static uint64_t bef_sky_padding(size_t inbyte,
 				uint16_t il_n, uint16_t k, uint64_t bsize)
 {
-	uint64_t common = k * il_n * 64;
+	uint64_t common = (uint32_t) k * il_n * 64;
 	uint64_t pbyte = il_n * bsize;
 
 	if(pbyte % common != 0)
